+++ CHANGELOG.md
@@ -5,7 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
-TODO
+### Added
+
+- `Timex.set/2` now also accepts setting the `:date` from a `%Date{}` struct.
 
 ## 3.4.1